refactor: move notifications feature to features folder structure
- Move NotificationsBell.tsx from components/layout/header to features/notifications
- Move NotificationsContext.tsx from contexts to features/notifications
- Create index.ts for centralized exports
- Update all import statements to use new feature location
- Follow feature-based folder organization pattern
