+++ src/contexts/AppContext.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+// TODO: This is a temporary context for the app. It will be replaced with a more robust context later, probably each feature will have its own context.
+
 import React, { createContext, useContext, useState, ReactNode } from "react";
 
 interface AppContextType {
