+++ src/features/notifications/index.ts
@@ -1,0 +1,3 @@
+export { NotificationsBell } from './NotificationsBell';
+export { NotificationsProvider, useNotifications } from './NotificationsContext';
+export type { Notification } from './NotificationsContext';
